Separating scraper and query builder
