+++ appendices/aa_old.csv
@@ -1,74 +1,70 @@
 ﻿,Location
-0,Colwood
-1,Langford
-2,Metchosin
-3,Sooke
-4,View Royal
-5,Esquimalt
-6,Juan de Fuca
-7,Oak Bay
-8,Saanich
-9,Salt Spring Island
-10,Sidney
-11,Victoria
+0,900-block Pandora Avenue
+1,Beacon Hill Park
+2,Burnside-Gorge
+3,Cecelia Ravine Park
+4,Central Park
+5,Central Saanich
+6,Coast Salish
+7,Colwood
+8,Cook Street Village
+9,Esquimalt
+10,Fairfield
+11,Fairfield-Gonzales
 12,Fernwood
-13,Gonzales
-14,Harris Green
-15,North Park
-16,Oaklands
-17,Rockland
-18,Uplands
-19,Burnside-Gorge
-20,Cook Street Village
-21,Fairfield-Gonzales
-22,Gonzales
-23,Foul Bay
-24,Hillside-Quadra
-25,Millstream
-26,Oak Bay
-27,Quadra
-28,Quadra Village
-29,Rock Bay
-30,Rockheights
-31,Saanichton
-32,Songhees
-33,The Gorge
-34,Vic West
-35,Victoria West
-36,West Shore
-37,Beacon Hill Park
-38,Central Park
-39,Gonzales Park
-40,Hollywood Park
-41,Irving Park
-42,Oaklands Park
-43,Royal Athletic Park
-44,Stadacona Park
-45,Topaz Park
-46,Cecelia Ravine Park
-47,Harris Green
-48,Point Ellice Park
+13,Galloping Goose
+14,Gonzales
+15,Foul Bay
+16,Harris Green
+17,Highlands
+18,Hillside-Quadra
+19,Hollywood Park
+20,Irving Park
+21,Juan de Fuca
+22,Kwakwaka’wakw
+23,Langford
+24,Lekwungen
+25,Lochside
+26,Malahat
+27,Metchosin
+28,Millstream
+29,North Park
+30,North Saanich
+31,Nuu-chah-nulth
+32,Nuu-chah-nulth
+33,Oak Bay
+34,Oaklands
+35,Oaklands Park
+36,Pacheedaht
+37,Pandora Avenue
+38,Pauquachin
+39,Point Ellice Park
+40,Quadra
+41,Rock Bay
+42,Rockheights
+43,Rockland
+44,Royal Athletic Park
+45,Saanich
+46,Saanichton
+47,Salt Spring Island
+48,Sc'ianew
 49,Selkirk Green
-50,Songhees Walkway
-51,Vic West Park
-52,Galloping Goose
-53,Selkirk Trestle
-54,Lochside
-55,900-block Pandora Avenue
-56,Pandora Avenue
-57,Coast Salish
-58,Nuu-chah-nulth
-59,Kwakwaka’wakw
-60,Lekwungen
-61,Songhees
-62,Esquimalt
-63,WSÁNEĆ
-64,Tsartlip
-65,Tsawout
-66,Pauquachin
-67,Tseycum
-68,Malahat
-69,Nuu-chah-nulth
-70,Sc'ianew
-71,T'Sou-ke
-72,Pacheedaht
+50,Selkirk Trestle
+51,Sidney
+52,Songhees
+53,Sooke
+54,Southern Gulf Islands
+55,Stadacona Park
+56,The Gorge
+57,Topaz Park
+58,Tsartlip
+59,Tsawout
+60,Tseycum
+61,T'Sou-ke
+62,Uplands
+63,Vic West
+64,Victoria
+65,Victoria West
+66,View Royal
+67,West Shore
+68,WSÁNEĆ
