+++ appendices/abpersons.csv
@@ -1,52 +1,52 @@
 ﻿,Influencers
-0,   Adam_Stirling 
-1,   CharlesBodi 
-2,   NicoleChaland 
-3,   Buckleygkml 
-4,   Darylbassett3 
-5,   sarahpottsvic 
-6,   eric_doherty 
-7,   self_govern4us 
-8,   murray_langdon 
-9,   dave_eby 
-10,   GraceALore 
-11,   Stephen_Andrew 
-12,   ninethreeseven 
-13,   coryresilient 
-14,   HattieCosta1 
-15,   photowarrior 
-16,   CarolinaIbarra 
-17,   CoreyRanger 
-18,   MarikaAlbert 
-19,   Laurel_BC 
-20,   zacdevries 
-21,   CouncllrNTaylor 
-22,   ChiefManak 
-23,   victoriasandy 
-24,   PACochran 
-25,   jfatkey 
-26,   KristaLoughton 
-27,   SongheesChief 
-28,   spayneBC 
-29,   DavidHScreech 
-30,   MurdochOakBay 
-31,   ColinPlant2018 
-32,   BarbDesjardins 
-33,   CTess1 
-34,   JeremyLoveday 
-35,   Louise_Hartland 
-36,   BarclayWallace 
-37,   Hilarylmarks 
-38,   Fairfield_Comm 
-39,   AdamPOlsen 
-40,   kevinalbersBC 
-41,   KatStinson 
-42,   chartj88 
-43,   BerniePauly 
-44,   shelliegudgeon 
-45,   BruceAlready 
-46,   MajaTait 
-47,   AdrianDix 
-48,   r_garrison 
-49,   ElizabethMay 
-50,   AlanRycroft 
+0,Adam_Stirling
+1,CharlesBodi
+2,NicoleChaland
+3,Buckleygkml
+4,Darylbassett3
+5,sarahpottsvic
+6,eric_doherty
+7,self_govern4us
+8,murray_langdon
+9,dave_eby
+10,GraceALore
+11,Stephen_Andrew
+12,ninethreeseven
+13,coryresilient
+14,HattieCosta1
+15,photowarrior
+16,CarolinaIbarra
+17,CoreyRanger
+18,MarikaAlbert
+19,Laurel_BC
+20,zacdevries
+21,CouncllrNTaylor
+22,ChiefManak
+23,victoriasandy
+24,PACochran
+25,jfatkey
+26,KristaLoughton
+27,SongheesChief
+28,spayneBC
+29,DavidHScreech
+30,MurdochOakBay
+31,ColinPlant2018
+32,BarbDesjardins
+33,CTess1
+34,JeremyLoveday
+35,Louise_Hartland
+36,BarclayWallace
+37,Hilarylmarks
+38,Fairfield_Comm
+39,AdamPOlsen
+40,kevinalbersBC
+41,KatStinson
+42,chartj88
+43,BerniePauly
+44,shelliegudgeon
+45,BruceAlready
+46,MajaTait
+47,AdrianDix
+48,r_garrison
+49,ElizabethMay
+50,AlanRycroft
